fixed delete backup_links

+++ renaissancebot/db/user_backup_accounts/remove_backup_acounts.py
@@ -1,5 +1,4 @@
 from datetime import datetime, timedelta
-
 import aiosqlite
 
 
@@ -10,25 +9,32 @@ async def remove_old_backup_accounts():
 
         # Текущая дата и время
         current_time = datetime.now()
+        cutoff_time = current_time - timedelta(days=1)
 
         # Проверяем связи, которым больше 24 часов
         await cursor.execute('''
-            SELECT user_id, email
+            SELECT user_id, email, link_date
             FROM UserBackupAccounts
             WHERE link_date <= ?
-        ''', (current_time - timedelta(days=1),))
+        ''', (cutoff_time,))
 
         # Получаем все записи, у которых истек срок действия
         rows = await cursor.fetchall()
 
-        for row in rows:
-            user_id, email = row
+        if rows:
+            for row in rows:
+                user_id, email, link_date = row
 
-            # Удаляем связь, если прошло более 24 часов
-            await cursor.execute('''
-                DELETE FROM UserBackupAccounts
-                WHERE user_id = ? AND email = ?
-            ''', (user_id, email))
+                # Выводим данные, которые собираемся удалить для отладки
+                print(f"Deleting user_id: {user_id}, email: {email}, link_date: {link_date}")
+
+                # Удаляем связь, если прошло более 24 часов
+                await cursor.execute('''
+                    DELETE FROM UserBackupAccounts
+                    WHERE user_id = ? AND email = ?
+                ''', (user_id, email))
+        else:
+            print("No rows to delete.")
 
         # Применяем изменения
         await conn.commit()
